Version 2
Made a LOT of changes. The BoxSlider, TiltWindow and ColorPickerTester
are the only classes I didn't modify.

List of changes:
- Made event serializable so that it can be used in the inspector
- Refactoring
- placed all components that are placed in the prefab into a UI folder
- The ColorPicker (previously HSVPicker) is now only a container of a
color. It does not modify anything itself. It does however send events
to all listeners when something changes one of the values (R, G, B, A,
H, S, V or the CurrentColor).
- All other elements (sliders, hex input field, etc) have a reference to
the ColorPicker (previously this was the other way around, but that got
messy once I added my stuff). They listen to the colorchanged event of
the ColorPicker and update their value, when their value changes they
also change the value of the ColorPicker
- Made a single slider script for each value (R, G, B, A, H, S, V).
(added the a, s and v sliders)
- Made a script that creates the background for a slider and keeps it up
to date if the color changes (the same way the Unity color picker works)
- Hue slider now works properly (no longer inverted)
- Hex field now updates roperly.
- New demo to display changes